cpumask: Add cpumasks for big and LITTLE CPU clusters
Add cpu_lp_mask and cpu_perf_mask to represent the CPUs that belong to
each cluster in a dual-cluster, heterogeneous system.

Now this has been modified for rosy, that is, homogeneous system.
The little cpu mask (In defconfig) is set to 240, which in binary is 11110000, referring to cpu7, cpu6, cpu5, cpu4, cpu3, cpu2, cpu1, cpu0.
And the big cpu mask (In defconfig) has been set to 15, which in binary is 1111, referring to cpu4, cpu5, cpu6, cpu7

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: I5e9425df3c537c1a209f12e010be36448bd053db